Filthy hack to prevent postfix DNS resolution err.
postfix runs in a chroot under /var/spool/postfix.  When it needs to
resolve the name of a machine to which it will relay mail, it performs
a DNS lookup, which uses /etc/resolv.conf from the chroot env.

If this is not present, the lookup will fail. We can't copy the root
file-system during Docker Compose, because the one that is needed
is generated shortly after boot.

Here we put the copy in the CMD. We do an nslookup first to be
totally sure the file is there and correct.